+++ config/routes.rb
@@ -1,6 +1,9 @@
 Rails.application.routes.draw do
 
   namespace :admin do
+    resources :partners
+  end
+  namespace :admin do
     resources :team_members
   end